pyproject.toml
[project]
name = "smchess"
version = "1.0.0"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.12.0",
    "pyright>=1.1.405",
    "python-chess>=1.999",
    "ruff>=0.13.0",
    "uvicorn>=0.35.0",
    "httpx>=0.25.0",
    "pillow>=10.0.0",
    "cairosvg>=2.7.0",
    "fastapi>=0.116.1",
    "stockfish>=3.28.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
